change types

+++ utils.h
@@ -1,5 +1,5 @@
 #pragma once
 #include <Arduino.h>
 
-float getTorque(float rpm);
+float getTorque(uint16_t rpm);
 int16_t getCoolantTemp(uint8_t AA);
